requirements.txt
streamlit
openai
pyaudio
SpeechRecognition
pyobjc